requirements.txt
asgiref==3.8.1
Django==5.0.4
sqlparse==0.5.0
tzdata==2024.1
django-allauth==0.62.1
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
